fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
lizard==1.17.25
mando==0.7.1
numpy==2.2.4
pandas==2.2.3
pathspec==0.12.1
pydantic==2.11.3
pydantic_core==2.33.1
PyDriller==2.7
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
radon==6.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
types-pytz==2025.2.0.20250326
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.1
